refactor: set up folder scructure and lightly applied mvc pattern
