+++ README.md
@@ -1,6 +1,6 @@
 # OPIc-magician-API
 
-v1.0.0
+Version 1.0.1
 
 ## 기능
